Added max height and width
