+++ public/index.html
@@ -7,6 +7,6 @@
 </head>
 
 <body>
-    <div id="sceneContainer"></div>
+    <div id="sceneContainer" style="height: 100%; width: 100%"></div>
 </body>
 </html>
